requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.12.0
django-jsonfield==1.4.1
django-redis==5.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
jsonfield==3.1.0
munch==2.5.0
numpy==1.22.4
packaging==21.3
pandas==1.4.2
psycopg2==2.9.3
pyparsing==3.0.9
pyproj==3.3.1
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
Shapely==1.8.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.1.0
wrapt==1.14.1
